v1/data/pyproject.toml
[tool.poetry]
name = "data"
version = "0.1.0"
description = ""
authors = ["Abhijit Gupta <avgupta456@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.21.5"
requests = "^2.26.0"
scipy = "^1.7.3"
SQLAlchemy = "^1.4.28"
sqlalchemy-cockroachdb = "^1.4.3"
python-dotenv = "^0.19.2"
psycopg2 = "^2.9.2"
attrs = "^21.2.0"
fastapi = "^0.71.0"
uvicorn = "^0.16.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pyinstrument = "^4.1.1"
psutil = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
